2020/02/17 Commit

+++ ABC_C/ABC138_C_Alchemist.cpp
@@ -19,7 +19,7 @@ double fnMaxValueGet(const vector<double>& cnrvnValue)
   {
     sort(begin(vnAscValue), end(vnAscValue));
     auto it = begin(vnAscValue);
-    double nAveValue = (double)( *it + *(it + 1) ) / 2;
+    double nAveValue = ( *it + *(it + 1) ) / 2;
 
     vnAscValue.erase(it, it + 2);
     vnAscValue.push_back(nAveValue);
